refactor(night-w3): split cluster C oversize files (OI-1085/1201/1205)
Split 3 files exceeding 800-line threshold into focused sub-modules:

- append_receipt.py (1272→414 lines): extracted receipt_cache.py,
  receipt_enrichment.py, receipt_git_session.py, receipt_quality_oi.py
- subprocess_dispatch.py (1154→682 lines): extracted dispatch_context.py,
  dispatch_git_ops.py, dispatch_git_query.py, dispatch_manifest.py,
  dispatch_receipt.py
- receipt_processor_v4.sh (1247→436 lines): extracted receipt_proc_helpers.sh,
  receipt_proc_sections.sh

All extracted modules ≤500 lines. Public API re-exported for backward compat.
Test patch targets updated to match new module locations.

Dispatch-ID: 20260430-night-w3-refactor-c
